kvm-unit-tests: configure changes for illumos.
Warn, don't fail, if the check for `getopt -T` fails in the `configure`
script.

Aside from this check, `configure` does not use `getopt`, so don't
fail to run if `getopt -T` doesn't indicate support for  the extended
Linux version.  Getopt is only used in `run_tests.sh`, which tests for
extended getopt anyway, but emit a warning here.

Signed-off-by: Dan Cross <cross@oxidecomputer.com>
Message-Id: <20220601215749.30223-2-cross@oxidecomputer.com>
Signed-off-by: Thomas Huth <thuth@redhat.com>