s390x: Introduce symbol for lowcore and use it
This gets rid of bunch of pointers pointing to the lowcore used in
various places and replaces it with a unified way of addressing the
lowcore.
The new symbol is not a pointer. While this will lead to worse code
generation (cannot use register 0 for addressing), that should not
matter too much for kvm unit tests.
Since the lowcore is located per definition at address 0, the symbol is
defined via the linker scripts.
The symbol also will be used to create pointers that the compiler cannot
warn about as being outside the bounds of an array.

Signed-off-by: Janis Schoetterl-Glausch <scgl@linux.ibm.com>
Reviewed-by: Thomas Huth <thuth@redhat.com>
Message-Id: <20220608122953.1051952-2-scgl@linux.ibm.com>
Signed-off-by: Claudio Imbrenda <imbrenda@linux.ibm.com>