ioapic: poll for completion of other CPUs in level-triggered SMP tests
The IOAPIC does not do anything if a level-triggered interrupt is asserted
while its REMOTE_IRR bit is set in the redirection table, and this can
cause a race condition between test_ioapic_physical_destination_mode
and test_ioapic_logical_destination_mode.  If CPU 1 is still running in
ioapic_isr_85 and has not sent the EOI, while CPU 0 has already reached
the set_irq_line(0x0e, 1) line of test_ioapic_logical_destination_mode,
the interrupt is dropped and the test hangs.

For some reason, this has started to occur about 20-30% of the time with
a 5.20-rc0 debug kernel.  With this patch, the ioapic test has survived
100 iterations without failing.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>