ci: Update the list of tests that we run in the Fedora Cirrus-CI
The "msr" test recently started failing since it got extended, but
the kernel of the CI machine is not new enough to handle the new
tests. Thus we have to disable it in the CI for now.
On the other hand, there are a couple of other tests which seem
to work now, so we can include those in the list instead.

Message-Id: <20220922122850.266424-1-thuth@redhat.com>
Signed-off-by: Thomas Huth <thuth@redhat.com>