x86/apic: Run tests that modify APIC ID and/or APIC_BASE after other tests
Run the sub-tests that modify the APIC ID and/or APIC base address after
all other sub-tests.  KVM disables APICv/AVIC if the ID or base is
changed, and keeps APICv/AVIC disabled even if the original value is
restored.  I.e. run as many tests as possible with an APICv/AVIC friendly
setup.

Signed-off-by: Sean Christopherson <seanjc@google.com>
Link: https://lore.kernel.org/r/20221001011301.2077437-8-seanjc@google.com