x86/apic: Add test for logical mode IPI delivery (cluster and flat)
Add an APIC sub-test to verify the darker corners of logical mode IPI
delivery.  Logical mode is rather bizarre, in that each "ID" is treated
as a bitmask, e.g. an ID with multiple bits set can match multiple
destinations.  Verify that overlapping and/or superfluous destinations
and IDs with multiple target vCPUs are handled correctly for both flat
and cluster modes.

Signed-off-by: Sean Christopherson <seanjc@google.com>
Link: https://lore.kernel.org/r/20221001011301.2077437-9-seanjc@google.com