gitlab-ci: Update to Fedora 37
Our gitlab-ci jobs were still running with Fedora 32 that is out of
service already. Let's update to Fedora 37 that brings a new QEMU
which also allows to run more tests with TCG. While we're at it,
also list each test in single lines and sort them alphabetically
so that it is easier to follow which tests get added and removed.
Beside adding some new tests, two entries are also removed here:
The "port80" test was removed a while ago from the x86 folder
already, but not from the .gitlab-ci.yml yet (seems like the run
script simply ignores unknown tests instead of complaining), and
the "tsc_adjust" is only skipping in the CI, so it's currently not
really usefull to try to run it in the CI.

Message-Id: <20221206104003.149630-1-thuth@redhat.com>
Acked-by: Claudio Imbrenda <imbrenda@linux.ibm.com>
Signed-off-by: Thomas Huth <thuth@redhat.com>