configure: Show the option in case it is not known
When mis-typing one of the options of the configure script, it shows
you the list of valid options, but does not tell you which option was
wrong. Then it can take a while until you figured out where the typo is.
Let's help the user here a little bit by printing which option had not
been understood.

Message-Id: <20230112095523.938919-1-thuth@redhat.com>
Reviewed-by: Philippe Mathieu-Daudé <philmd@linaro.org>
Signed-off-by: Thomas Huth <thuth@redhat.com>