arm/psci: Add PSCI CPU_OFF test case
The test uses the following method.

The primary CPU brings up all the secondary CPUs, which are held in a wait
loop. Once the primary releases the CPUs, each of the secondary CPUs
proceed to issue CPU_OFF.

The primary CPU then checks for the status of the individual CPU_OFF
request. There is a chance that some CPUs might return from the CPU_OFF
function call after the primary CPU has finished the scan. There is no
foolproof method to handle this, but the test tries its best to
eliminate these false positives by introducing an extra delay if all the
CPUs are reported offline after the initial scan.

[ Alex E: Skip CPU_OFF test if cpu onlining failed, drop cpu_off_success in
          favour of checking AFFINITY_INFO, commit message tweaking ]

Signed-off-by: Nikita Venkatesh <Nikita.Venkatesh@arm.com>
Signed-off-by: Alexandru Elisei <alexandru.elisei@arm.com>
Signed-off-by: Andrew Jones <andrew.jones@linux.dev>