x86: Run the tsc test with -cpu max
Run the tsc test with -cpu max to enable RDPID testing. The current
configuration (-cpu kvm64,+rdtscp) does not include RDPID even if the
host supports. In effect, test_rdpid() never runs.

Use max instead of adding +rdpid to the existing configuration to avoid
this biting someone in the future if they add a test for a new feature.

Fixes: 10631a5bebd8 ("x86: tsc: add rdpid test")
Signed-off-by: David Matlack <dmatlack@google.com>
Link: https://lore.kernel.org/r/20230307005547.607353-2-dmatlack@google.com
Signed-off-by: Sean Christopherson <seanjc@google.com>